refactor: extract unwrapJsonRpcResult into separate function in renderer process
