+++ packages/renderer-process/src/parts/UnwrapJsonRpcResult/UnwrapJsonRpcResult.js
@@ -1,0 +1,13 @@
+import { JsonRpcError } from '../JsonRpcError/JsonRpcError.js'
+import * as RestoreJsonRpcError from '../RestoreJsonRpcError/RestoreJsonRpcError.js'
+
+export const unwrapJsonRpcResult = (responseMessage) => {
+  if ('error' in responseMessage) {
+    const restoredError = RestoreJsonRpcError.restoreJsonRpcError(responseMessage.error)
+    throw restoredError
+  }
+  if ('result' in responseMessage) {
+    return responseMessage.result
+  }
+  throw new JsonRpcError('unexpected response message')
+}
